+++ CHANGELOG.md
@@ -4,5 +4,7 @@
 * Aug 17, 2023: New challenge: Chess Simp — "Poor Pawn Got Blundered 10 Times"
 * Aug 17, 2023: New challenge: Chess Simp — "A Check Will Ruin Everything"
 * Aug 18, 2023: "Open leaderboard" palette command
+* Aug 19, 2023: Ability to check if the challenge is explicitly lost
+* Aug 19, 2023: New challenge: Chess Simp — "This Is A Bit Too Aggressive"
 
 # Aug 17, 2023 — started the changelog